Ajustando após avaliação do PR

+++ SME.Acessos.Aplicacao/Servicos/ServicoTokenJwt.cs
@@ -17,7 +17,7 @@ namespace SME.Acessos.Aplicacao.Servicos
             this.jwtTokenSettings = jwtTokenSettings?.Value;
         }
 
-        public string GerarToken(string usuarioLogin, string usuarioNome, Guid guidPerfil, IEnumerable<int> permissionamentos)
+        public string GerarToken(string usuarioLogin, string usuarioNome, Guid guidPerfil, IEnumerable<long> permissionamentos)
         {
             List<Claim> claims = new List<Claim>();
